// photovoltaicV2/src/app/services/models/system-fixing.ts
/* tslint:disable */
/* eslint-disable */
export interface SystemFixing {
  adaptability?: string;
  charge?: number;
  company?: number;
  height?: number;
  id?: number;
  image?: string;
  installation_method?: string;
  materiel?: string;
  model?: string;
  price?: number;
  quantity?: number;
  type?: string;
  width?: number;
}